requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0

# Data & Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Export
onnx>=1.14.0
onnxruntime>=1.16.0

# Config & Utils
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.0.0

# Development
pytest>=7.4.0
httpx>=0.25.0